Main Extract

See merge request gmv-bda/upm/inesdata-mov/data-generation!14

+++ inesdata_mov_datasets/sources/aemet/extract/extract.py
@@ -1,20 +1,24 @@
 """Gather raw data from aemet."""
+
 import datetime
 import io
 import json
+from pathlib import Path
+from minio import Minio
 
 import requests
 
 from inesdata_mov_datasets.settings import Settings
-from inesdata_mov_datasets.utils import minio_connection, read_settings
+from inesdata_mov_datasets.utils import read_settings, check_minio_file_exists, check_local_file_exists
 
 
-def get_weather_madrid(config: Settings):
+def get_aemet(config: Settings, minio_client: Minio=None):
     """Request aemet API to get data from Madrid weather.
 
     Args:
         config (Settings): Object with the config file.
     """
+    print("EXTRACTING AEMET")
     url_madrid = (
         "https://opendata.aemet.es/opendata/api/prediccion/especifica/municipio/horaria/28079"
     )
@@ -28,12 +32,13 @@ def get_weather_madrid(config: Settings):
     r = requests.get(url_madrid, headers=headers)
     r_json = requests.get(r.json()["datos"]).json()
 
-    save_aemet(config, r_json)
+    save_aemet(config, r_json, minio_client)
+    
+    print("EXTRACTED AEMET")
+    print("- - - - - - -")
 
-    return
 
-
-def save_aemet(config: Settings, data: json):
+def save_aemet(config: Settings, data: json, minio_client: Minio=None):
     """Save weather json.
 
     Args:
@@ -48,22 +53,37 @@ def save_aemet(config: Settings, data: json):
         "%Y/%m/%d"
     )  # formatted date year/month/day for storage in Minio
 
-    if config.storage.default == "minio":
-        minio_client = minio_connection(config)
-        object_name = (
-            f"/raw/aemet/{formatted_date_slash}/madrid/madrid_weather_{formatted_date_day}.json"
-        )
-        response_json_str = json.dumps(data)
-        minio_client.put_object(
-            config.storage.config.minio.bucket,
-            object_name,
-            io.BytesIO(response_json_str.encode("utf-8")),
-            len(response_json_str),
-        )
+    if config.storage.default == "minio":        
+        # Define object name
+        object_name = Path("raw") / "aemet" / formatted_date_slash / f"aemet_{formatted_date_day}.json"
+        
+        if not check_minio_file_exists(minio_client, config.storage.config.minio.bucket, str(object_name)):
+            # Convert data to JSON string
+            response_json_str = json.dumps(data)
+            
+            # Create BytesIO object from JSON string
+            response_bytes = io.BytesIO(response_json_str.encode("utf-8"))
+            
+            # Put object in Minio bucket
+            minio_client.put_object(
+                config.storage.config.minio.bucket,
+                str(object_name),
+                response_bytes,
+                len(response_json_str)
+            )
 
-    return
-
-
-if __name__ == "__main__":
-    config = read_settings("/home/jfog/proyects/inesdata-mov/data-generation/jfog-newconfig.yaml")
-    get_weather_madrid(config)
+    if config.storage.default == "local":
+        # Define object name and path for local storage
+        object_name = f"aemet_{formatted_date_day}.json"
+        local_path = Path(config.storage.config.local.path) / "raw" / "aemet" / formatted_date_slash
+        
+        if not check_local_file_exists(local_path, object_name):
+            # Convert data to JSON string
+            response_json_str = json.dumps(data)
+            
+            # Create directories if they don't exist
+            local_path.mkdir(parents=True, exist_ok=True)
+            
+            # Write JSON data to file
+            with open(local_path / object_name, "w") as file:
+                file.write(response_json_str)
